requirements.txt
Django==4.2.27
Pillow==11.3.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
gunicorn==21.2.0
dj-database-url==2.1.0
